Recreated Login as TaggedMutationResolver
